+++ src/components/homePage/itemVideo/ItemVideo.module.css
@@ -4,9 +4,10 @@
 }
 
 .itemVideo {
-    width: 70vw;
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
     border-radius: 2vw;
-    margin-top: 5vh;
 }
 
 .loopingText {
@@ -23,4 +24,29 @@
 
 .loopingTextEven {
     color: var(--main-background-color);
+}
+
+.coveredVideo {
+    width: 70vw;
+    height: 70vh;
+    position: relative;
+    top: 0;
+    left: 0;
+    margin-top: 5vh;
+}
+
+.arrow {
+    height: 10%;
+    position: absolute;
+    top: 45%;
+}
+
+.arrowLeft {
+    left: 0;
+    transform: rotate(90deg);
+}
+
+.arrowRight {
+    right: 0;
+    transform: rotate(-90deg);
 }
