+++ src/components/homePage/orderForm/OrderCreation.module.css
@@ -1,7 +1,7 @@
 .orderCreation {
     justify-content: space-between;
     width: 60vw;
-    height: 75vh;
+    height: 90vh;
 }
 
 .infoContainer {
