+++ src/components/homePage/orderForm/inputs/Input.module.css
@@ -68,7 +68,7 @@
 }
 
 .listInput {
-    width: 80%;
+    width: 100%;
     margin-right: 5%;
     border: none;
     background-color: transparent;
